Adding a starting state as argument
I also gave to policy gradient (and visit_dist) an extra argument which is the starting state. Actually it should be a starting distribution. I think that it is ok to get different results for different starting states (but maybe not in this example where you can transition easily from one state to another).